Add External object class interface
External objects need to be identified in the restriction results in a means
that is understandable to the external system.  This class provides a common
identifier type.  Currently, it is a simple unsigned long and could be extended
in future if need be.

The constructor for External is deliberately protected to force that is only
instantiated as an inherited type, never standalone.  It's only use is for
representing other classes externally.