Allow adding child model Nodes to each other
One Model Node can be created as a child of another.  This allows a hierarchy
to be constructed for our tree of filters in the Asset Allocation.

Deliberately chose to store the child nodes of a Node as a set of pointers,
rather than directely.  Whilst it would be more efficient to store the objects
directly in the vector, adding children would invalidate references and pointers
when the vector expands.