AssetClassFilter implementation
Adds a filtering class for Assets based on a specific filter the Assets do or
do not have.  This will allow the most common type of filtering of Assets as
part of financial models.

Note: sic::Asset::hasClass was made virtual to allow GMock to mock it in unit
testing.  This was necessary to segregate the AssetClassFilterTest unit test
from the Asset implementation.