Use boost any_iterator to allow generic Iterators
A difficulty that came up in testing was producing loosely-coupled APIs for
sources of items.  This was due to generics requiring concrete iterator classes
to be passed.

Through type-elusion, boost's any_iterator class allows us to pass any iterator
(set, vector, ...) and store it in Iterators.  In later testing, we only need to
pass this Iterators / Source class.